requirements.txt
alive_progress==2.4.1
nltk==3.7
pyenchant==3.2.2
